+++ NEWS.md
@@ -1,4 +1,4 @@
-# rgl  0.107.19
+# rgl  0.107.20
 
 ## Major changes
 
@@ -8,6 +8,10 @@ will be used as a "mipmap".  This may be used in
 of the display.
 * Added `getBoundary3d()` function to extract the boundary
 edges of a mesh.
+* Added material property `user_id`, a string associated 
+with each object.  The value is reported by `ids3d()` and
+may be used to select objects in `pop3d()`, but otherwise is 
+largely ignored by `rgl`.
 
 ## Minor changes
 * Allowed for a third coordinate in `text3d()`'s `adj`